gles2: introduce egl::context
The egl::context class is used to create the context and surface
on which we'll draw ou grames using OpenGL ES 2. It's not tied
to a specific version of OpenGL ES.

The code in this commit is still quite crude, and will need to
be adjusted. The TDO file lists some shortcomings that will be
addressed soon.