app: introduce on_create()
on_create() emulates the call to ANativeActivity_onCreate() in
the context of the app thread. It feeds the user app with the
app_activity instance.

This behavior replaces the behavior of on_start() which was
used to feed the user app with the activity object (this was
not very logical).