external: adding libpng-1.5.14 to the source tree
There must be a better way to do that, but since libpng is not part of
the official Android distribution, I had to do something. Note that
we must compile this as a static library - and we cannot rely on
the configure script (it's likely to not work very well in our
specific case).